+++ order-query-ms/src/main/java/ibm/labs/kc/order/query/dao/OrderDAOMock.java
@@ -11,8 +11,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ibm.labs.kc.order.query.model.QueryOrder;
-
 
 public class OrderDAOMock implements OrderDAO {
     private static final Logger logger = LoggerFactory.getLogger(OrderDAOMock.class);
