+++ order-query-ms/src/main/java/ibm/labs/kc/order/query/model/Order.java
@@ -1,6 +1,7 @@
 package ibm.labs.kc.order.query.model;
 
-import java.util.Objects;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 
 public class Order {
 
@@ -95,11 +96,11 @@ public class Order {
 
     @Override
     public int hashCode() {
-        return Objects.hash(this);
+        return HashCodeBuilder.reflectionHashCode(this);
     }
 
     @Override
     public boolean equals(Object obj) {
-        return Objects.equals(this, obj);
+        return EqualsBuilder.reflectionEquals(this, obj);
     }
 }
